+++ src/app/layout.tsx
@@ -1,26 +1,27 @@
-import DeployButton from "@/src/components/deploy-button";
-import { EnvVarWarning } from "@/src/components/env-var-warning";
-import HeaderAuth from "@/src/components/header-auth";
-import { ThemeSwitcher } from "@/src/components/theme-switcher";
-import { hasEnvVars } from "@/src/utils/supabase/check-env-vars";
-import { Geist } from "next/font/google";
-import { ThemeProvider } from "next-themes";
-import Link from "next/link";
-import "./globals.css";
+import DeployButton from '@/components/deploy-button';
+import { EnvVarWarning } from '@/components/env-var-warning';
+import HeaderAuth from '@/components/header-auth';
+import { ThemeSwitcher } from '@/components/theme-switcher';
+import { hasEnvVars } from '@/utils/supabase/check-env-vars';
+import { Geist } from 'next/font/google';
+import { ThemeProvider } from 'next-themes';
+import Link from 'next/link';
+import './globals.css';
+import { Toaster } from '@/components/ui/toaster';
 
 const defaultUrl = process.env.VERCEL_URL
   ? `https://${process.env.VERCEL_URL}`
-  : "http://localhost:3000";
+  : 'http://localhost:3000';
 
 export const metadata = {
   metadataBase: new URL(defaultUrl),
-  title: "Next.js and Supabase Starter Kit",
-  description: "The fastest way to build apps with Next.js and Supabase",
+  title: 'Next.js and Supabase Starter Kit',
+  description: 'The fastest way to build apps with Next.js and Supabase',
 };
 
 const geistSans = Geist({
-  display: "swap",
-  subsets: ["latin"],
+  display: 'swap',
+  subsets: ['latin'],
 });
 
 export default function RootLayout({
@@ -42,7 +43,7 @@ export default function RootLayout({
               <nav className="w-full flex justify-center border-b border-b-foreground/10 h-16">
                 <div className="w-full max-w-5xl flex justify-between items-center p-3 px-5 text-sm">
                   <div className="flex gap-5 items-center font-semibold">
-                    <Link href={"/"}>Next.js Supabase Starter</Link>
+                    <Link href={'/'}>Next.js Supabase Starter</Link>
                     <div className="flex items-center gap-2">
                       <DeployButton />
                     </div>
@@ -50,13 +51,11 @@ export default function RootLayout({
                   {!hasEnvVars ? <EnvVarWarning /> : <HeaderAuth />}
                 </div>
               </nav>
-              <div className="flex flex-col gap-20 max-w-5xl p-5">
-                {children}
-              </div>
+              <div className="flex flex-col gap-20 max-w-5xl p-5">{children}</div>
 
               <footer className="w-full flex items-center justify-center border-t mx-auto text-center text-xs gap-8 py-16">
                 <p>
-                  Powered by{" "}
+                  Powered by{' '}
                   <a
                     href="https://supabase.com/?utm_source=create-next-app&utm_medium=template&utm_term=nextjs"
                     target="_blank"
@@ -71,6 +70,7 @@ export default function RootLayout({
             </div>
           </main>
         </ThemeProvider>
+        <Toaster />
       </body>
     </html>
   );