+++ src/app/page.tsx
@@ -1,7 +1,7 @@
-import Hero from "@/src/components/hero";
-import ConnectSupabaseSteps from "@/src/components/tutorial/connect-supabase-steps";
-import SignUpUserSteps from "@/src/components/tutorial/sign-up-user-steps";
-import { hasEnvVars } from "@/src/utils/supabase/check-env-vars";
+import Hero from '@/components/hero';
+import ConnectSupabaseSteps from '@/components/tutorial/connect-supabase-steps';
+import SignUpUserSteps from '@/components/tutorial/sign-up-user-steps';
+import { hasEnvVars } from '@/utils/supabase/check-env-vars';
 
 export default async function Home() {
   return (