+++ src/components/header-auth.tsx
@@ -1,9 +1,10 @@
-import { signOutAction } from "@/src/app/actions";
-import { hasEnvVars } from "@/src/utils/supabase/check-env-vars";
-import Link from "next/link";
-import { Badge } from "./ui/badge";
-import { Button } from "./ui/button";
-import { createClient } from "@/src/utils/supabase/server";
+import { signOutAction } from '@/app/actions';
+import { hasEnvVars } from '@/utils/supabase/check-env-vars';
+import Link from 'next/link';
+import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/ui/button';
+import { createClient } from '@/utils/supabase/server';
+import { AuthModal } from './auth/auth-modal';
 
 export default async function AuthButton() {
   const supabase = await createClient();
@@ -17,10 +18,7 @@ export default async function AuthButton() {
       <>
         <div className="flex gap-4 items-center">
           <div>
-            <Badge
-              variant={"default"}
-              className="font-normal pointer-events-none"
-            >
+            <Badge variant={'default'} className="font-normal pointer-events-none">
               Please update .env.local file with anon key and url
             </Badge>
           </div>
@@ -28,7 +26,7 @@ export default async function AuthButton() {
             <Button
               asChild
               size="sm"
-              variant={"outline"}
+              variant={'outline'}
               disabled
               className="opacity-75 cursor-none pointer-events-none"
             >
@@ -37,7 +35,7 @@ export default async function AuthButton() {
             <Button
               asChild
               size="sm"
-              variant={"default"}
+              variant={'default'}
               disabled
               className="opacity-75 cursor-none pointer-events-none"
             >
@@ -52,19 +50,14 @@ export default async function AuthButton() {
     <div className="flex items-center gap-4">
       Hey, {user.email}!
       <form action={signOutAction}>
-        <Button type="submit" variant={"outline"}>
+        <Button type="submit" variant={'outline'}>
           Sign out
         </Button>
       </form>
     </div>
   ) : (
     <div className="flex gap-2">
-      <Button asChild size="sm" variant={"outline"}>
-        <Link href="/sign-in">Sign in</Link>
-      </Button>
-      <Button asChild size="sm" variant={"default"}>
-        <Link href="/sign-up">Sign up</Link>
-      </Button>
+      <AuthModal />
     </div>
   );
 }