+++ src/lib/utils.ts
@@ -1,6 +1,11 @@
 import { clsx, type ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
 
+/**
+ * Merges Tailwind CSS class names conditionally.
+ * @param inputs - List of class values.
+ * @returns A single string of merged class names.
+ */
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }